+++ finalProject/src/main/java/com/univ/fin/member/model/dao/MemberDao.java
@@ -58,19 +58,24 @@ public class MemberDao {
 		return (ArrayList)sqlSession.selectList("memberMapper.majorClass", departmentName);
 	}
 	
-	// 수강신청 - 강의시간표 -> 학년도,학기 조회
+	// 강의시간표 -> 학년도,학기 조회
 	public ArrayList<String> selectClassTerm(SqlSessionTemplate sqlSession) {
 		return (ArrayList)sqlSession.selectList("memberMapper.selectClassTerm");
 	}
 	
-	// 수강신청 - 강의시간표 -> 단과대학별 전공 조회
+	// 강의시간표 -> 단과대학별 전공 조회
 	public ArrayList<String> selectDepartment(SqlSessionTemplate sqlSession, String college) {
 		return (ArrayList)sqlSession.selectList("memberMapper.selectDepartment", college);
 	}
 
-	// 수강신청 - 강의시간표 -> 전공 선택 후 전공수업 조회
+	// 강의시간표 -> 전공 선택 후 전공수업 조회
 	public ArrayList<Classes> selectDepartmentMajor(SqlSessionTemplate sqlSession, HashMap<String, String> map) {
 		return (ArrayList)sqlSession.selectList("memberMapper.selectDepartmentMajor", map);
+	}
+
+	// 강의시간표 -> 교양수업 조회
+	public ArrayList<Classes> selectElective(SqlSessionTemplate sqlSession, HashMap<String, String> map) {
+		return (ArrayList)sqlSession.selectList("memberMapper.selectElective", map);
 	}
