+++ finalProject/src/main/java/com/univ/fin/member/model/service/MemberService.java
@@ -36,14 +36,17 @@ public interface MemberService {
 	//수강신청 - 학부전공별 조회
 	ArrayList<Classes> majorClass(String departmentName);
 	
-	// 수강신청 - 강의시간표 -> 학년도,학기 조회
+	// 강의시간표 -> 학년도,학기 조회
 	ArrayList<String> selectClassTerm();
 	
-	// 수강신청 - 강의시간표 -> 단과대학별 전공 조회
+	// 강의시간표 -> 단과대학별 전공 조회
 	ArrayList<String> selectDepertment(String college);
 
-	// 수강신청 - 강의시간표 -> 전공 선택 후 전공수업 조회
+	// 강의시간표 -> 전공 선택 후 전공수업 조회
 	ArrayList<Classes> selectDepartmentMajor(HashMap<String, String> map);
+
+	// 강의시간표 -> 교양수업 조회
+	ArrayList<Classes> selectElective(HashMap<String, String> map);
